Let DocumentSidebar make use of the renamed PrePublish component (now AnalysisChecklist).
