+++ js/src/containers/DocumentSidebar.js
@@ -1,25 +1,29 @@
 import { withSelect, withDispatch } from "@wordpress/data";
 import { compose } from "@wordpress/compose";
 import getIndicatorForScore from "../analysis/getIndicatorForScore";
-import DocumentSidebar from "../components/DocumentSidebar";
+import AnalysisChecklist from "../components/AnalysisChecklist";
+
+import {
+	addReadabilityCheck,
+	addSEOCheck,
+	addSchemaBlocksValidationCheck,
+} from "../helpers/addCheckToChecklist";
 
 export default compose( [
 	withSelect( ( select ) => {
 		const data = select( "yoast-seo/editor" );
-		const focusKeyphrase = data.getFocusKeyphrase();
 		const seoScoreIndicator = getIndicatorForScore( data.getResultsForFocusKeyword().overallScore );
 		const readabilityScoreIndicator = getIndicatorForScore( data.getReadabilityResults().overallScore );
 		const { isKeywordAnalysisActive, isContentAnalysisActive } = data.getPreferences();
+		const schemaBlocksValidationResults = data.getSchemaBlocksValidationResults();
 
-		return {
-			focusKeyphrase,
-			isKeywordAnalysisActive,
-			isContentAnalysisActive,
-			seoScore: seoScoreIndicator.className,
-			seoScoreLabel: seoScoreIndicator.screenReaderReadabilityText,
-			readabilityScore: readabilityScoreIndicator.className,
-			readabilityScoreLabel: readabilityScoreIndicator.screenReaderReadabilityText,
-		};
+		const checklist = [];
+
+		addReadabilityCheck( checklist, readabilityScoreIndicator, isKeywordAnalysisActive );
+		addSEOCheck( checklist, seoScoreIndicator, isContentAnalysisActive );
+		addSchemaBlocksValidationCheck( checklist, schemaBlocksValidationResults );
+
+		return { checklist, shouldShowIntro: false };
 	} ),
 	withDispatch( ( dispatch ) => {
 		const { openGeneralSidebar } = dispatch(
@@ -36,4 +40,4 @@ export default compose( [
 
 		return { onClick };
 	} ),
-] )( DocumentSidebar );
+] )( AnalysisChecklist );
